Clean up serializers in views.py
