+++ freva/urls.py
@@ -25,5 +25,3 @@ urlpatterns = [
     path("api/", include(("runs.urls", "runs"), namespace="runs")),
     path("history/", include(("history.urls", "history"), namespace="history")),
 ]
-
-print(urlpatterns)
